Fix logic for keyboard support
My charNotAllowed varibale initially was setup with a value like:
"...JKLMNOPQRSTUVWXYZTabCapslockControl...".

My if statement checked to see if the key pressed (e.key) was included
in the charNotAllowed variable.  If it was, then it would be ignored.
Since some keys are full words (ie "CapsLock") it was preventing keys
such as "C" from working (which I used for clearing the screen).  I made
the variable otherKeysNotAllowed to filter out keys that are full words.